apacheSubProjectJavaRefactored/test/java/org/apache/beam/sdk/transforms/windowing/TriggerTest.java
package org.apache.beam.sdk.transforms.windowing;

import static org.junit.Assert.assertEquals;
import static org.hamcrest.MatcherAssert.assertThat; // Optimized by LLM: Suggestion 3 applied
import static org.hamcrest.Matchers.is; // Optimized by LLM: Suggestion 3 applied
import static org.hamcrest.Matchers.equalTo; // Optimized by LLM: Suggestion 3 applied
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import java.util.Arrays;
import java.util.Collections; // Optimized by LLM: Suggestion 2 applied
import java.util.List;
import org.joda.time.Instant;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.junit.runners.JUnit4;

@RunWith(JUnit4.class)
public class TriggerTest {

  @Test
  public void testTriggerToString() {
    assertEquals("AfterWatermark.pastEndOfWindow()", AfterWatermark.pastEndOfWindow().toString());
    assertEquals(
        "Repeatedly.forever(AfterWatermark.pastEndOfWindow())",
        Repeatedly.forever(AfterWatermark.pastEndOfWindow()).toString());
  }

  @Test
  public void testIsCompatible() {
    assertThat(new Trigger1(Collections.emptyList()).isCompatible(new Trigger1(Collections.emptyList())), is(true)); // Optimized by LLM: Suggestion 1 and 2 applied
    assertThat(
        new Trigger1(Arrays.asList(new Trigger2(Collections.emptyList())))
            .isCompatible(new Trigger1(Arrays.asList(new Trigger2(Collections.emptyList())))), is(true)); // Optimized by LLM: Suggestion 1 and 2 applied

    assertThat(new Trigger1(Collections.emptyList()).isCompatible(new Trigger2(Collections.emptyList())), is(false)); // Optimized by LLM: Suggestion 1 and 2 applied
    assertThat(
        new Trigger1(Arrays.asList(new Trigger1(Collections.emptyList())))
            .isCompatible(new Trigger1(Arrays.asList(new Trigger2(Collections.emptyList())))), is(false)); // Optimized by LLM: Suggestion 1 and 2 applied
  }

  private static class Trigger1 extends Trigger {

    private Trigger1(List<Trigger> subTriggers) {
      super(subTriggers);
    }

    @Override
    protected Trigger getContinuationTrigger(List<Trigger> continuationTriggers) {
      return null;
    }

    @Override
    public Instant getWatermarkThatGuaranteesFiring(BoundedWindow window) {
      return null;
    }

    @Override
    public boolean mayFinish() {
      return false;
    }
  }

  private static class Trigger2 extends Trigger {

    private Trigger2(List<Trigger> subTriggers) {
      super(subTriggers);
    }

    @Override
    protected Trigger getContinuationTrigger(List<Trigger> continuationTriggers) {
      return null;
    }

    @Override
    public Instant getWatermarkThatGuaranteesFiring(BoundedWindow window) {
      return null;
    }

    @Override
    public boolean mayFinish() {
      return false;
    }
  }
}